+++ packages/components/box/docs.mdx
@@ -1,0 +1,86 @@
+---
+name: Box
+route: /components/box
+menu: Components
+---
+
+import { Playground, Props } from "docz"
+import Box from './index'
+import { Flex } from "@chakra-ui/core"
+
+# Box
+
+Box component
+
+The HTML DOM structure is
+
+```
+  <div></div>
+```
+
+
+
+## Usage
+<Playground>
+  <Flex as="form" p="1">
+    <Box
+      as="input"
+      placeholder="Your email"
+      type="email"
+      flex="1"
+      py={2}
+      px={4}
+      rounded="md"
+      bg="gray.100"
+      borderWidth="1px"
+      _hover={{ borderColor: "gray.200", bg: "gray.200" }}
+      _focus={{
+        outline: "none",
+        bg: "white",
+        boxShadow: "outline",
+        borderColor: "gray.300",
+      }}
+    />
+    <Box
+      as="button"
+      bg="teal.500"
+      py={2}
+      px={4}
+      ml={3}
+      rounded="md"
+      fontWeight="semibold"
+      color="white"
+      _hover={{ bg: "teal.600" }}
+      _focus={{ boxShadow: "outline" }}
+    >
+      Sign Up
+    </Box>
+  </Flex>
+</Playground>
+
+___
+
+## Hover
+
+Add the _hover prop to only apply style props on hover.
+
+<Playground>
+  <Box
+    as="button"
+    color="blue.700"
+    fontWeight="semibold"
+    py={2}
+    px={4}
+    borderWidth="1px"
+    borderColor="blue.500"
+    rounded="md"
+    _hover={{ bg: "blue.500", color: " white" }}
+    _focus={{ boxShadow: "outline" }}
+  >
+    Hover me
+  </Box>
+</Playground>
+
+## Props
+
+<Props of={Box}/>
