+++ packages/components/flex/docs.mdx
@@ -1,0 +1,93 @@
+---
+name: Flex
+route: /components/flex
+menu: Components
+---
+
+import { Playground, Props } from "docz"
+import Flex from './index'
+import Box from "../box"
+
+# Box
+
+Box component
+
+The HTML DOM structure is
+
+```
+  <div></div>
+```
+
+
+
+## Usage
+<Playground>
+  <Flex as="form" p="1">
+    <Box
+      as="button"
+      bg="teal.500"
+      py={2}
+      px={4}
+      rounded="md"
+      fontWeight="semibold"
+      color="white"
+      _hover={{ bg: "teal.600" }}
+      _focus={{ boxShadow: "outline" }}
+    >
+      Button
+    </Box>
+    <Box
+      as="button"
+      bg="blue.500"
+      py={2}
+      px={4}
+      ml={2}
+      rounded="md"
+      fontWeight="semibold"
+      color="white"
+      _hover={{ bg: "teal.600" }}
+      _focus={{ boxShadow: "outline" }}
+    >
+      Button
+    </Box>
+  </Flex>
+</Playground>
+
+
+## Usage
+<Playground>
+  <Flex.column as="form" p="1">
+    <Box
+      as="button"
+      width="min-content"
+      bg="teal.500"
+      py={2}
+      px={4}
+      rounded="md"
+      fontWeight="semibold"
+      color="white"
+      _hover={{ bg: "teal.600" }}
+      _focus={{ boxShadow: "outline" }}
+    >
+      Button
+    </Box>
+    <Box
+      as="button"
+      width="min-content"
+      bg="blue.500"
+      py={2}
+      px={4}
+      rounded="md"
+      fontWeight="semibold"
+      color="white"
+      _hover={{ bg: "teal.600" }}
+      _focus={{ boxShadow: "outline" }}
+    >
+      Button
+    </Box>
+  </Flex.column>
+</Playground>
+
+## Props
+
+<Props of={Flex}/>
